Continued server-side logic
There are still some errors when trying to send backups to the server. The build script was changed so that it now creates separate directories for each client and server with it's own jar file, keystore and truststore. Because of the addition of a build script and the changes it makes to the run process the README also needs to be changed